fix: move select logic into select dir
